+++ src/internal/mod.rs
@@ -8,12 +8,12 @@ pub(crate) use character_class_registry::CharacterClassRegistry;
 
 /// Module that provides functions and types related to compiled Lookahead.
 pub(crate) mod compiled_lookahead;
-// pub(crate) use compiled_lookahead::{CompiledLookahead, CompiledNfaLookahead};
+// pub(crate) use compiled_lookahead::{CompiledLookahead, CompiledDfaLookahead};
 pub(crate) use compiled_lookahead::CompiledLookahead;
 
 /// Module that provides functions and types related to compiled NFA.
-pub(crate) mod compiled_nfa;
-// pub(crate) use compiled_nfa::CompiledNfa;
+pub(crate) mod compiled_dfa;
+// pub(crate) use compiled_dfa::CompiledDfa;
 
 /// Module that provides functions and types related to compiled ScannerModes.
 pub(crate) mod compiled_scanner_mode;
@@ -38,6 +38,9 @@ pub(crate) use ids::{
 pub(crate) mod match_function;
 pub(crate) use match_function::MatchFunction;
 
+/// Module that provides functions and types related to DFA minimization.
+pub(crate) mod minimizer;
+
 /// Module that provides functions and types related to the multi pattern NFA.
 pub(crate) mod multi_pattern_nfa;
 pub(crate) use multi_pattern_nfa::MultiPatternNfa;
@@ -54,5 +57,5 @@ mod scanner_cache;
 pub(crate) use scanner_cache::SCANNER_CACHE;
 
 /// Module that provides functions and types related to NFA scanner implementations.
-pub(crate) mod scanner_nfa_impl;
-pub(crate) use scanner_nfa_impl::ScannerNfaImpl;
+pub(crate) mod scanner_impl;
+pub(crate) use scanner_impl::ScannerImpl;