+++ src/internal/scanner_cache.rs
@@ -1,6 +1,6 @@
 //! Module with the global scanner cache.
 //!
-//! The scanner cache is a global cache that stores `ScannerNfaImpl`` instances. It is used to
+//! The scanner cache is a global cache that stores `ScannerImpl`` instances. It is used to
 //! create scanners with the same scanner modes without having to recompile the regular
 //! expressions for the patterns and than generate the NFA for the scanner.
 //!
@@ -15,11 +15,11 @@ use rustc_hash::FxHashMap;
 
 use std::sync::Arc;
 
-use super::ScannerNfaImpl;
+use super::ScannerImpl;
 
-/// The cache is a `FxHashMap` that maps a vectors of `ScannerMode` to `Arc<ScannerNfaImpl>`.
+/// The cache is a `FxHashMap` that maps a vectors of `ScannerMode` to `Arc<ScannerImpl>`.
 pub(crate) struct ScannerCache {
-    cache: FxHashMap<Vec<ScannerMode>, Arc<ScannerNfaImpl>>,
+    cache: FxHashMap<Vec<ScannerMode>, Arc<ScannerImpl>>,
 }
 
 impl ScannerCache {
@@ -36,14 +36,14 @@ impl ScannerCache {
     /// This function uses `unsafe` because it dereferences a raw pointer.
     /// The Arc assures that the pointer is valid because it holds a reference to the object as
     /// long as the Arc is in the cache and thus alive.
-    pub(crate) fn get(&mut self, modes: &[ScannerMode]) -> Result<ScannerNfaImpl> {
+    pub(crate) fn get(&mut self, modes: &[ScannerMode]) -> Result<ScannerImpl> {
         if let Some(scanner) = self.cache.get(modes) {
             // We need to clone the scanner because we need to return a new instance of the scanner.
             // This is because the scanner is mutable and we need to have a unique instance of the
             // scanner.
             // But cloning is much cheaper than creating a new scanner.
             let cloned_scanner =
-                unsafe { (*std::sync::Arc::<ScannerNfaImpl>::as_ptr(scanner)).clone() };
+                unsafe { (*std::sync::Arc::<ScannerImpl>::as_ptr(scanner)).clone() };
             Ok(cloned_scanner)
         } else {
             self.cache